Подключает функции из отдельного файла
